separate function for sorting
